backend/backend/requirements.txt
Django>=4.2
django-cors-headers
python-dotenv
requests
dj-database-url
psycopg[binary]
